fix(client-api-context): infinite loop + Fast Refresh split
GraphistryScene's ref callback depended on the full ctx object, which
gets a new reference every time Provider's value memo recomputes (i.e.,
every setRpc). In React 19 that triggered the ref-reattach cycle:
setRef regenerates → React detaches old ref → calls old(null) →
registerIframe(null) → setRpc(null, dispose) → re-render → loop until
"Maximum update depth exceeded". Pulling the stable pieces
(registerIframe, sceneSrc) out of ctx makes setRef's identity stable.

Also splits the module so Fast Refresh works when editing context.tsx:

- internal.ts: Ctx object + shared types (ControlledDomain,
  GraphistryHandle, InternalContext).
- context.tsx: GraphistryProvider + GraphistryScene components only.
- hooks.ts: useGraphistry, useGraphistryScene, useSelection, useLabels,
  useFilters — all in one hooks file.

React Fast Refresh bails on mixed component + non-component exports,
forcing full reloads on every tsx edit. With this split, tsx edits
hot-patch without tearing down the iframe + handshake state.

example/App.tsx: default HOST to `${location.hostname}:8491` so the
app works out-of-box against Alex's dev Graphistry instead of the
public hub (which runs pre-Protocol-v2 viz code). Override with
?host=… query param.
